requirements.txt
Flask
pyserial

# CNN 및 ML 관련 의존성
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
Pillow>=10.0.0

# 지속적 학습 스케줄러
schedule>=1.2.0